feat: Add layout cubit and screens
This commit adds a layout cubit and screens for the Discover and Translate screens. It also includes a new AppRoot widget and updates to the main.dart file to use it. Additionally, new utility classes for app settings and navigation were added. The pubspec.yaml file was also updated to include new dependencies.